Remove Diagnostic Source + Configuration Manager from .NET Standard DLL as it's not supported (#535)
