+++ tools/specs/Microsoft.Data.SqlClient.nuspec
@@ -60,7 +60,6 @@ When using NuGet 3.x this package requires at least version 3.4.</description>
         <dependency id="runtime.native.System.Data.SqlClient.sni" version="4.4.0" exclude="Compile" />
         <dependency id="Microsoft.Win32.Registry" version="4.7.0" exclude="Compile" />
         <dependency id="System.Buffers" version="4.5.1" exclude="Compile" />
-        <dependency id="System.Diagnostics.DiagnosticSource" version="4.7.0" exclude="Compile" />
         <dependency id="System.Memory" version="4.5.4" exclude="Compile" />
         <dependency id="System.Security.Principal.Windows" version="4.7.0" exclude="Compile" />
         <dependency id="System.Text.Encoding.CodePages" version="4.7.0" exclude="Compile" />
